requirements.txt
Appium-Python-Client==2.0.0
pytest==6.2.5
pytest-base-url==1.4.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-selenium==2.0.1
pytest-variables==1.9.0
pytest-xdist==2.4.0
requests==2.26.0
selenium==4.0.0